+++ packages/core/lib/combo/confvars.ts
@@ -205,6 +205,7 @@ export const CONFVARS = [
   'MM_PRE_ACTIVATED_OWL_ZC',
   'MM_PRE_ACTIVATED_OWL_IC',
   'MM_PRE_ACTIVATED_OWL_ST',
+  'MM_WELL_OPEN',
 ] as const;
 
 export type Confvar = typeof CONFVARS[number];