+++ packages/core/lib/combo/patch-build/randomizer.ts
@@ -1008,6 +1008,7 @@ function worldConfig(world: World, settings: Settings): Set<Confvar> {
     MM_PRE_ACTIVATED_OWL_ZC: world.resolvedFlags.mmPreActivatedOwls.has('zoracape'),
     MM_PRE_ACTIVATED_OWL_IC: world.resolvedFlags.mmPreActivatedOwls.has('canyon'),
     MM_PRE_ACTIVATED_OWL_ST: world.resolvedFlags.mmPreActivatedOwls.has('tower'),
+    MM_WELL_OPEN: settings.beneathWell === 'open',
   };
 
   for (const v in exprs) {