+++ packages/core/lib/combo/settings/data.ts
@@ -848,6 +848,19 @@ export const SETTINGS = [{
   default: 'none',
   cond: hasMM,
 }, {
+  key: 'beneathWell',
+  name: 'Beneath the Well behavior (MM)',
+  category: 'main.events',
+  type: 'enum',
+  description: 'Controls which MM dungeons have ',
+  values: [
+    { value: 'vanilla', name: 'Vanilla', description: 'Nothing changes, Gibdos are back once you reset the cycle.' },
+    { value: 'remorseless', name: 'Remorseless', description: 'Once you gave a Gibdo its item, it is permanently gone, also affects the ending ladder and the sun block.' },
+    { value: 'open', name: 'Open', description: 'All the Gibdos are already gone, also applies to the ending ladder and the sun block.' }
+  ],
+  default: 'vanilla',
+  cond: hasMM,
+}, {
   key: 'kakarikoGate',
   name: 'Kakariko Gate',
   category: 'main.events',